+++ src/device_builder.rs
@@ -8,6 +8,7 @@ pub struct UsbVidPid(pub u16, pub u16);
 /// Used to build new [`UsbDevice`]s.
 pub struct UsbDeviceBuilder<'a, B: UsbBus> {
     alloc: &'a UsbBusAllocator<B>,
+    control_buffer: &'a mut [u8],
     config: Config<'a>,
 }
 
@@ -32,6 +33,8 @@ pub enum BuilderError {
     InvalidPacketSize,
     /// Configuration specifies higher USB power draw than allowed
     PowerTooHigh,
+    /// The provided control buffer is too small for the provided maximum packet size.
+    ControlBufferTooSmall,
 }
 
 /// Provides basic string descriptors about the device, including the manufacturer, product name,
@@ -82,9 +85,14 @@ impl<'a> StringDescriptors<'a> {
 
 impl<'a, B: UsbBus> UsbDeviceBuilder<'a, B> {
     /// Creates a builder for constructing a new [`UsbDevice`].
-    pub fn new(alloc: &'a UsbBusAllocator<B>, vid_pid: UsbVidPid) -> UsbDeviceBuilder<'a, B> {
+    pub fn new(
+        alloc: &'a UsbBusAllocator<B>,
+        vid_pid: UsbVidPid,
+        control_buffer: &'a mut [u8],
+    ) -> UsbDeviceBuilder<'a, B> {
         UsbDeviceBuilder {
             alloc,
+            control_buffer,
             config: Config {
                 device_class: 0x00,
                 device_sub_class: 0x00,
@@ -104,8 +112,16 @@ impl<'a, B: UsbBus> UsbDeviceBuilder<'a, B> {
     }
 
     /// Creates the [`UsbDevice`] instance with the configuration in this builder.
-    pub fn build(self) -> UsbDevice<'a, B> {
-        UsbDevice::build(self.alloc, self.config)
+    pub fn build(self) -> Result<UsbDevice<'a, B>, BuilderError> {
+        if self.control_buffer.len() < self.config.max_packet_size_0 as usize {
+            return Err(BuilderError::ControlBufferTooSmall);
+        }
+
+        Ok(UsbDevice::build(
+            self.alloc,
+            self.config,
+            self.control_buffer,
+        ))
     }
 
     builder_fields! {
@@ -186,6 +202,10 @@ impl<'a, B: UsbBus> UsbDeviceBuilder<'a, B> {
         match max_packet_size_0 {
             8 | 16 | 32 | 64 => {}
             _ => return Err(BuilderError::InvalidPacketSize),
+        }
+
+        if self.control_buffer.len() < max_packet_size_0 as usize {
+            return Err(BuilderError::ControlBufferTooSmall);
         }
 
         self.config.max_packet_size_0 = max_packet_size_0;