+++ src/test_class.rs
@@ -1,9 +1,9 @@
 #![allow(missing_docs)]
 
 use crate::class_prelude::*;
-use crate::descriptor::lang_id::LangID;
 use crate::device::{StringDescriptors, UsbDevice, UsbDeviceBuilder, UsbVidPid};
 use crate::Result;
+use core::cell::UnsafeCell;
 use core::cmp;
 
 #[cfg(feature = "test-class-high-speed")]
@@ -21,6 +21,8 @@ mod sizes {
     pub const BULK_ENDPOINT: u16 = 64;
     pub const INTERRUPT_ENDPOINT: u16 = 31;
 }
+
+static mut CONTROL_BUFFER: UnsafeCell<[u8; 256]> = UnsafeCell::new([0; 256]);
 
 /// Test USB class for testing USB driver implementations. Supports various endpoint types and
 /// requests for testing USB peripheral drivers on actual hardware.
@@ -94,7 +96,7 @@ impl<B: UsbBus> TestClass<'_, B> {
 
     /// Convenience method to create a UsbDevice that is configured correctly for TestClass.
     pub fn make_device<'a>(&self, usb_bus: &'a UsbBusAllocator<B>) -> UsbDevice<'a, B> {
-        self.make_device_builder(usb_bus).build()
+        self.make_device_builder(usb_bus).build().unwrap()
     }
 
     /// Convenience method to create a UsbDeviceBuilder that is configured correctly for TestClass.
@@ -112,14 +114,16 @@ impl<B: UsbBus> TestClass<'_, B> {
         &self,
         usb_bus: &'a UsbBusAllocator<B>,
     ) -> UsbDeviceBuilder<'a, B> {
-        UsbDeviceBuilder::new(usb_bus, UsbVidPid(VID, PID))
-            .strings(&[StringDescriptors::default()
-                .manufacturer(MANUFACTURER)
-                .product(PRODUCT)
-                .serial_number(SERIAL_NUMBER)])
-            .unwrap()
-            .max_packet_size_0(sizes::CONTROL_ENDPOINT)
-            .unwrap()
+        UsbDeviceBuilder::new(usb_bus, UsbVidPid(VID, PID), unsafe {
+            CONTROL_BUFFER.get_mut()
+        })
+        .strings(&[StringDescriptors::default()
+            .manufacturer(MANUFACTURER)
+            .product(PRODUCT)
+            .serial_number(SERIAL_NUMBER)])
+        .unwrap()
+        .max_packet_size_0(sizes::CONTROL_ENDPOINT)
+        .unwrap()
     }
 
     /// Must be called after polling the UsbDevice.